requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
albumentations>=1.3.1
Pillow>=9.5.0

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Utilities
tqdm>=4.65.0
pathlib

# Optional but Recommended
jupyter>=1.0.0
ipykernel>=6.22.0

# For Kaggle datasets (optional)
kaggle>=1.5.16

# For production deployment (optional)
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0